pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cdss-agentic-rag"
version = "0.1.0"
description = "Clinical Decision Support System with Agentic RAG on Azure"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "CDSS Team" },
]
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.9",

    # Azure OpenAI
    "openai>=1.60.0",

    # Azure Services
    "azure-search-documents>=11.6.0",
    "azure-cosmos>=4.9.0",
    "azure-ai-documentintelligence>=1.0.0",
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.9.0",
    "azure-storage-blob>=12.24.0",

    # PubMed / Entrez
    "biopython>=1.84",

    # Async HTTP client for drug APIs
    "httpx>=0.28.0",

    # Caching
    "redis[hiredis]>=5.2.0",

    # Embeddings / numerical
    "numpy>=2.0.0",

    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cdss"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["cdss"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
